Modify structure and call chain
First part of a great refactor

examples/example2.c

add linux/version.h include (for KERNEL_VERSION macro)
move mem_sampling_backed structure here (from include/numap.h)
moved msb_lock here (from src/numap.c)
moved global pointer msb here (from src/numap.c)
add handler function (args measure and fd)
	-> does the data backup
moved free_mem_sampling_backed here (from src/numap.c)
moved numap_sampling_print_backed here (from src/numap_analyse.c)
changed call to numap_sampling_set_mode_buffer_flush and isolate it
removed temp_debug_print and replaced it by direct call of
	numap_sampling_print_backed and free_mem_sampling_backed

include/numap.h

moved link_fd_measure (to src/numap.c)
moved mem_sampling_backed (to examples/example2/c)
add handler, missed and nb_refresh fields to numap_sampling_measure
removed declarations
	-> temp_debug_print
	-> numap_sampling_print_backed
changed signature of numap_sampling_set_mode_buffer_flush

src/numap.c

moved msb_lock (to examples/example2)
moved link_fd_measure here (from include/numap.h)
moved msb (to examples/example2.c)
indentation fixes
replaced all data backup part by call of a user defined handler
	-> if no handler set, counts missed measures
moved free_mem_sampling_backed (to examples/example2.c)
changed signature of numap_sampling_set_mode_buffer_flush
	-> now takes a user-defined handler that receives a measure
		pointer and a file descriptor
	-> this handler is called by the library handler
removed temp_debug_print

src/numap_analyse.c

moved numap_sampling_print_backed (to examples/example2.c)
