+++ src/numap_analyse.c
@@ -203,86 +203,6 @@ int get_index(uint32_t tid, struct numap_sampling_measure *measure) {
   return -1;
 }
 
-int numap_sampling_print_backed(struct numap_sampling_measure *measure, struct mem_sampling_backed *msb, char print_samples) {
-  int thread;
-  if (msb == NULL) return 0;
-  for (thread = 0; thread < measure->nb_threads; thread++) {
-    struct mem_sampling_backed* current_msb = msb;
-    int fd = measure->fd_per_tid[thread];
-
-    struct perf_event_header *header;
-    uint64_t consumed = 0;
-    int na_miss_count = 0;
-    int cache1_count = 0;
-    int cache2_count = 0;
-    int cache3_count = 0;
-    int lfb_count = 0;
-    int memory_count = 0;
-    int remote_memory_count = 0;
-    int remote_cache_count = 0;
-    int total_count = 0;
-    while (current_msb != NULL) {
-      header  = (struct perf_event_header *)(current_msb->buffer);
-      int head = 0;
-      if (current_msb->fd == fd) {
-        while (head < current_msb->buffer_size) {
-          if (header->size == 0) {
-            fprintf(stderr, "Error: invalid header size = 0\n");
-            return -1;
-          }
-          if (header -> type == PERF_RECORD_SAMPLE) {
-            struct sample *sample = (struct sample *)((char *)(header) + 8);
-            if (is_served_by_local_NA_miss(sample->data_src)) {
-              na_miss_count++;
-            }
-            if (is_served_by_local_cache1(sample->data_src)) {
-              cache1_count++;
-            }
-            if (is_served_by_local_cache2(sample->data_src)) {
-              cache2_count++;
-            }
-            if (is_served_by_local_cache3(sample->data_src)) {
-              cache3_count++;
-            }
-            if (is_served_by_local_lfb(sample->data_src)) {
-              lfb_count++;
-            }
-              if (is_served_by_local_memory(sample->data_src)) {
-              memory_count++;
-            }
-            if (is_served_by_remote_memory(sample->data_src)) {
-              remote_memory_count++;
-            }
-            if (is_served_by_remote_cache_or_local_memory(sample->data_src)) {
-              remote_cache_count++;
-            }
-            total_count++;
-            if (print_samples) {
-              printf("pc=%" PRIx64 ", @=%" PRIx64 ", src level=%s, latency=%" PRIu64 "\n", sample->ip, sample->addr, get_data_src_level(sample->data_src), sample->weight);
-            }
-          }
-          head += header->size;
-          header = (struct perf_event_header *)((char*)current_msb->buffer+head);
-        }
-      }
-      current_msb = current_msb->next;
-    }
-    printf("\n");
-    //printf("head = %" PRIu64 " compared to max = %zu\n", head, measure->mmap_len);
-    printf("Thread %d: %-8d samples\n", thread, total_count);
-    printf("Thread %d: %-8d %-30s %0.3f%%\n", thread, cache1_count, "local cache 1", (100.0 * cache1_count / total_count));
-    printf("Thread %d: %-8d %-30s %0.3f%%\n", thread, cache2_count, "local cache 2", (100.0 * cache2_count / total_count));
-    printf("Thread %d: %-8d %-30s %0.3f%%\n", thread, cache3_count, "local cache 3", (100.0 * cache3_count / total_count));
-    printf("Thread %d: %-8d %-30s %0.3f%%\n", thread, lfb_count, "local cache LFB", (100.0 * lfb_count / total_count));
-    printf("Thread %d: %-8d %-30s %0.3f%%\n", thread, memory_count, "local memory", (100.0 * memory_count / total_count));
-    printf("Thread %d: %-8d %-30s %0.3f%%\n", thread, remote_cache_count, "remote cache or local memory", (100.0 * remote_cache_count / total_count));
-    printf("Thread %d: %-8d %-30s %0.3f%%\n", thread, remote_memory_count, "remote memory", (100.0 * remote_memory_count / total_count));
-    printf("Thread %d: %-8d %-30s %0.3f%%\n", thread, na_miss_count, "unknown l3 miss", (100.0 * na_miss_count / total_count));
-  }
-
-  return 0;
-}
-
 int numap_sampling_print(struct numap_sampling_measure *measure, char print_samples) {
   int thread;
   for (thread = 0; thread < measure->nb_threads; thread++) {
